+++ pyproject.toml
@@ -127,8 +127,6 @@ per-file-ignores = [
   "cli/core/handlers/excel_file_handler_mixins.py:WPS210",
   "cli/core/handlers/horizontal_tab_file_manager.py:WPS210",
   "cli/core/mpt/api.py:WPS210",
-  "cli/core/price_lists/app/export.py:WPS210,WPS213,WPS231",
-  "cli/core/price_lists/app/sync.py:WPS210,WPS231",
   "cli/core/price_lists/services/item_service.py:WPS210,WPS231",
   "cli/core/products/app.py:WPS210,WPS213,WPS231",
   "cli/core/products/models/product.py:WPS210",
