+++ tests/cli/core/price_lists/app/test_sync.py
@@ -1,0 +1,220 @@
+import re
+
+from cli.core.price_lists import app
+from cli.core.price_lists.services import ItemService, PriceListService
+from cli.core.services.service_result import ServiceResult
+from cli.core.stats import PriceListStatsCollector
+from typer.testing import CliRunner
+
+runner = CliRunner()
+
+
+def strip_ansi(text):
+    return re.sub(r"\x1b\[[0-9;]*m", "", text)
+
+
+def test_sync_price_lists_not_files_found(price_list_new_file):
+    result = runner.invoke(app, ["sync", "some-file.xlsx"])
+
+    assert result.exit_code == 3, result.stdout
+    assert "No files found for provided paths" in strip_ansi(result.stdout)
+
+
+def test_sync_price_lists_multiple_files(
+    mocker, price_list_data_from_json, price_list_new_file, active_vendor_account
+):
+    mocker.patch(
+        "cli.core.price_lists.app.sync.get_active_account",
+        return_value=active_vendor_account,
+        autospec=True,
+    )
+    mocker.patch(
+        "cli.core.price_lists.app.sync.get_files_path",
+        return_value=[price_list_new_file, price_list_new_file],
+        autospec=True,
+    )
+    stats = PriceListStatsCollector()
+    price_list_service_retrieve_mock = mocker.patch.object(
+        PriceListService,
+        "retrieve",
+        return_value=ServiceResult(success=True, model=None, stats=stats),
+    )
+    price_list_service_create_mock = mocker.patch.object(
+        PriceListService,
+        "create",
+        return_value=ServiceResult(success=True, model=price_list_data_from_json, stats=stats),
+    )
+    item_service_update_mock = mocker.patch.object(
+        ItemService,
+        "update",
+        return_value=ServiceResult(success=True, model=None, stats=stats),
+    )
+
+    result = runner.invoke(
+        app,
+        ["sync", "fake_dir_path"],
+        input="y\ny\ny\ny\n",
+    )
+
+    assert result.exit_code == 0, result.stdout
+    assert price_list_service_retrieve_mock.call_count == 2
+    assert price_list_service_create_mock.call_count == 2
+    assert item_service_update_mock.call_count == 2
+
+
+def test_sync_price_lists_create(
+    mocker,
+    price_list_data_from_json,
+    price_list_file_path,
+    active_vendor_account,
+):
+    mocker.patch(
+        "cli.core.price_lists.app.sync.get_active_account",
+        return_value=active_vendor_account,
+        autospec=True,
+    )
+    stats = PriceListStatsCollector()
+    price_list_service_retrieve_mock = mocker.patch.object(
+        PriceListService,
+        "retrieve",
+        return_value=ServiceResult(success=True, model=None, stats=stats),
+    )
+    price_list_service_create_mock = mocker.patch.object(
+        PriceListService,
+        "create",
+        return_value=ServiceResult(success=True, model=price_list_data_from_json, stats=stats),
+    )
+    item_service_update_mock = mocker.patch.object(
+        ItemService,
+        "update",
+        return_value=ServiceResult(success=True, model=None, stats=stats),
+    )
+    price_list_service_update_spy = mocker.spy(PriceListService, "update")
+
+    result = runner.invoke(app, ["sync", str(price_list_file_path)], input="y\ny\n")
+
+    assert result.exit_code == 0, result.stdout
+    price_list_service_retrieve_mock.assert_called_once()
+    price_list_service_create_mock.assert_called_once()
+    item_service_update_mock.assert_called_once()
+    price_list_service_update_spy.assert_not_called()
+
+
+def test_sync_price_lists_create_error(
+    mocker, active_operations_account, price_list_data_from_json, price_list_file_path
+):
+    mocker.patch(
+        "cli.core.price_lists.app.sync.get_active_account",
+        return_value=active_operations_account,
+    )
+    stats = PriceListStatsCollector()
+    price_list_service_retrieve_mock = mocker.patch.object(
+        PriceListService,
+        "retrieve",
+        return_value=ServiceResult(success=True, model=None, stats=stats),
+    )
+    price_list_service_create_mock = mocker.patch.object(
+        PriceListService,
+        "create",
+        return_value=ServiceResult(success=False, model=None, stats=stats),
+    )
+    item_service_update_spy = mocker.spy(ItemService, "update")
+
+    result = runner.invoke(app, ["sync", str(price_list_file_path)], input="y\ny\n")
+
+    assert result.exit_code == 4
+    assert "Price list sync FAILED\n" in strip_ansi(result.stdout)
+    price_list_service_retrieve_mock.assert_called_once()
+    price_list_service_create_mock.assert_called_once()
+    item_service_update_spy.assert_not_called()
+
+
+def test_sync_price_lists_retrieve_error(mocker, active_vendor_account, price_list_file_path):
+    mocker.patch(
+        "cli.core.price_lists.app.sync.get_active_account",
+        return_value=active_vendor_account,
+        autospec=True,
+    )
+    stats = PriceListStatsCollector()
+    price_list_service_retrieve_mock = mocker.patch.object(
+        PriceListService,
+        "retrieve",
+        return_value=ServiceResult(success=False, model=None, stats=stats),
+    )
+    price_list_service_create_spy = mocker.spy(PriceListService, "create")
+    price_list_service_update_spy = mocker.spy(PriceListService, "update")
+    item_service_update_spy = mocker.spy(ItemService, "update")
+
+    result = runner.invoke(app, ["sync", str(price_list_file_path)], input="y\n")
+
+    assert result.exit_code == 4
+    assert "Price list sync FAILED\n" in strip_ansi(result.stdout)
+    price_list_service_retrieve_mock.assert_called_once()
+    price_list_service_create_spy.assert_not_called()
+    price_list_service_update_spy.assert_not_called()
+    item_service_update_spy.assert_not_called()
+
+
+def test_sync_price_lists_update(
+    mocker, price_list_data_from_json, price_list_file_path, active_vendor_account
+):
+    stats = PriceListStatsCollector()
+    mocker.patch(
+        "cli.core.price_lists.app.sync.get_active_account",
+        return_value=active_vendor_account,
+        autospec=True,
+    )
+    price_list_service_retrieve_mock = mocker.patch.object(
+        PriceListService,
+        "retrieve",
+        return_value=ServiceResult(success=True, model=price_list_data_from_json, stats=stats),
+    )
+    price_list_service_update_mock = mocker.patch.object(
+        PriceListService,
+        "update",
+        return_value=ServiceResult(success=True, model=price_list_data_from_json, stats=stats),
+    )
+    item_service_update_mock = mocker.patch.object(
+        ItemService,
+        "update",
+        return_value=ServiceResult(success=True, model=None, stats=stats),
+    )
+    price_list_service_create_spy = mocker.spy(PriceListService, "create")
+
+    result = runner.invoke(app, ["sync", str(price_list_file_path)], input="y\ny\n")
+
+    assert result.exit_code == 0, result.stdout
+    price_list_service_retrieve_mock.assert_called_once()
+    price_list_service_update_mock.assert_called_once_with()
+    item_service_update_mock.assert_called_once()
+    price_list_service_create_spy.assert_not_called()
+
+
+def test_sync_price_lists_update_error(
+    mocker, active_vendor_account, price_list_data_from_json, price_list_file_path
+):
+    mocker.patch(
+        "cli.core.price_lists.app.sync.get_active_account",
+        return_value=active_vendor_account,
+        autospec=True,
+    )
+    stats = PriceListStatsCollector()
+    price_list_service_retrieve_mock = mocker.patch.object(
+        PriceListService,
+        "retrieve",
+        return_value=ServiceResult(success=True, model=price_list_data_from_json, stats=stats),
+    )
+    price_list_service_update_mock = mocker.patch.object(
+        PriceListService,
+        "update",
+        return_value=ServiceResult(success=False, model=None, stats=stats),
+    )
+    item_service_update_spy = mocker.spy(ItemService, "update")
+
+    result = runner.invoke(app, ["sync", str(price_list_file_path)], input="y\ny\n")
+
+    assert result.exit_code == 4
+    assert "Price list sync FAILED\n" in strip_ansi(result.stdout)
+    price_list_service_retrieve_mock.assert_called_once()
+    price_list_service_update_mock.assert_called_once()
+    item_service_update_spy.assert_not_called()
